+++ toolbox/fdc3-workbench/src/store/SystemLogStore.ts
@@ -10,6 +10,8 @@ import snackbarStore from './SnackbarStore.js';
 export type logMessagesName =
   | 'getFdc3'
   | 'getInfo'
+  | 'setInstanceMetadata'
+  | 'updateTitle'
   | 'getChannels'
   | 'getCurrentChannel'
   | 'joinUserChannel'